refactor: split agent studio quick actions
